started to modify the ocr service and looking into upgrading to handle new response type
