+++ src/services/index.ts
@@ -1,4 +1,6 @@
 export { isBaseApiError, BaseApiClientError } from './api/base/error';
 export { useTagScanStatus } from './notification/TagScanStatus';
 
-export { OCR } from './ocr/ocr';
+export { OCRValidator, AzureOCRValidator } from './ocr/validator';
+export { AzureOCRv2 } from './ocr/ocr';
+export { AzureOCRv4 } from './ocr/ocrRead';
