+++ src/services/ocr/ocrRead.ts
@@ -1,0 +1,9 @@
+import { AzureOCRv2 } from './ocr';
+
+/**
+ * - Change resources to new endpoint
+ * - Change response handling into handling the new response.
+ * - More?
+ */
+
+export class AzureOCRv4 extends AzureOCRv2 {}
