+++ src/types/camera.ts
@@ -2,6 +2,7 @@ import { TagSummaryDto } from '@equinor/echo-search';
 import { DeviceInformation } from '@equinor/echo-utils';
 
 import { OCRPayload, ParsedComputerVisionResponse } from '@types';
+import { OCRValidator } from '@services';
 
 export interface ScannerProps extends CameraProps {
   ocrService: OCRService;
